+++ projects/pastanaga-angular/src/lib/dropdown/dropdown.component.html
@@ -1,13 +1,12 @@
-@if (isDisplayed) {
-  <div
-    class="pa-popup"
-    [ngStyle]="style">
-    <div class="pa-popup-wrapper">
-      <ul
-        class="pa-menu"
-        [attr.role]="role">
-        <ng-content></ng-content>
-      </ul>
-    </div>
+<div
+  class="pa-popup"
+  [hidden]="!isDisplayed"
+  [ngStyle]="style">
+  <div class="pa-popup-wrapper">
+    <ul
+      class="pa-menu"
+      [attr.role]="role">
+      <ng-content></ng-content>
+    </ul>
   </div>
-}
+</div>
